remove api module
